+++ lib/lightning_web/live/workflow_live.ex
@@ -2,30 +2,34 @@ defmodule LightningWeb.WorkflowLive do
   @moduledoc false
   use LightningWeb, :live_view
 
-  on_mount({LightningWeb.Hooks, :project_scope})
+  on_mount {LightningWeb.Hooks, :project_scope}
 
   alias Lightning.Workflows
+  import LightningWeb.WorkflowLive.Components
 
   @impl true
   def render(assigns) do
+    assigns = assigns |> assign_new(:show_canvas, fn -> true end)
+
     ~H"""
     <Layout.page_content>
       <:header>
         <Layout.header title={@page_title} socket={@socket}>
           <.link navigate={
-            Routes.project_workflow_path(@socket, :new_job, @project.id)
+            Routes.project_job_index_path(@socket, :index, @project.id)
           }>
-            <Common.button>
+            <div class="inline-flex justify-center py-2 px-4 border border-transparent shadow-sm text-sm font-medium rounded-md text-white bg-secondary-200 hover:bg-secondary-300 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-secondary-500">
               <div class="h-full">
-                <Heroicons.plus class="h-4 w-4 inline-block" />
-                <span class="inline-block align-middle">New Job</span>
+                <Heroicons.table_cells solid class="h-4 w-4 inline-block" />
               </div>
-            </Common.button>
+            </div>
           </.link>
         </Layout.header>
       </:header>
       <div class="relative h-full">
         <%= case @live_action do %>
+          <% :index -> %>
+            <.workflow_list page={@page} project={@project} />
           <% :new_job -> %>
             <div class="absolute w-1/3 inset-y-0 right-0 bottom-0 z-10">
               <div
@@ -42,7 +46,12 @@ defmodule LightningWeb.WorkflowLive do
                   current_user={@current_user}
                   builder_state={@builder_state}
                   return_to={
-                    Routes.project_workflow_path(@socket, :show, @project.id)
+                    Routes.project_workflow_path(
+                      @socket,
+                      :show,
+                      @project.id,
+                      @current_workflow.id
+                    )
                   }
                 />
               </div>
@@ -58,24 +67,57 @@ defmodule LightningWeb.WorkflowLive do
                   current_user={@current_user}
                   builder_state={@builder_state}
                   return_to={
-                    Routes.project_workflow_path(@socket, :show, @project.id)
+                    Routes.project_workflow_path(
+                      @socket,
+                      :show,
+                      @project.id,
+                      @current_workflow.id
+                    )
                   }
                 />
               </div>
             </div>
+          <% :show -> %>
+            <%= if length(@current_workflow.jobs) == 0 do %>
+              <div class="w-1/2 h-16 text-center my-16 mx-auto pt-4">
+                <div class="font-semibold text-gray-500 pb-2">
+                  No Job
+                </div>
+                <div class="text-xs text-gray-400">
+                  <.link patch={
+                    Routes.project_workflow_path(
+                      @socket,
+                      :new_job,
+                      @project.id,
+                      @current_workflow.id
+                    )
+                  }>
+                    <Common.button>
+                      <div class="h-full">
+                        <Heroicons.plus class="h-4 w-4 inline-block" />
+                        <span class="inline-block align-middle">
+                          Add your first job
+                        </span>
+                      </div>
+                    </Common.button>
+                  </.link>
+                </div>
+              </div>
+            <% end %>
           <% :edit_workflow -> %>
             <div class="absolute top-0 right-0 m-2 z-10">
               <div class="w-80 bg-white rounded-md shadow-xl ring-1 ring-black ring-opacity-5 p-3">
                 <.live_component
                   module={LightningWeb.WorkflowLive.WorkflowInspector}
-                  id={@workflow.id}
-                  workflow={@workflow}
+                  id={@current_workflow.id}
+                  workflow={@current_workflow}
                   project={@project}
                   return_to={
                     Routes.project_workflow_path(
                       @socket,
                       :show,
-                      @project.id
+                      @project.id,
+                      @current_workflow.id
                     )
                   }
                 />
@@ -83,22 +125,35 @@ defmodule LightningWeb.WorkflowLive do
             </div>
           <% _ -> %>
         <% end %>
-        <div
-          phx-hook="WorkflowDiagram"
-          class="h-full w-full"
-          id={"hook-#{@project.id}"}
-          phx-update="ignore"
-          base-path={Routes.project_workflow_path(@socket, :show, @project.id)}
-          data-project-space={@encoded_project_space}
-        >
-        </div>
+        <%= if @show_canvas do %>
+          <div
+            phx-hook="WorkflowDiagram"
+            class="h-full w-full"
+            id={"hook-#{@project.id}"}
+            phx-update="ignore"
+            base-path={
+              Routes.project_workflow_path(
+                @socket,
+                :show,
+                @project.id,
+                @current_workflow.id
+              )
+            }
+            data-project-space={@encoded_project_space}
+          >
+          </div>
+        <% end %>
       </div>
     </Layout.page_content>
     """
   end
 
-  defp encode_project_space(project) do
-    Workflows.get_workflows_for(project)
+  defp encode_project_space(%Workflows.Workflow{} = workflow) do
+    workflow
+    |> Lightning.Repo.preload(
+      jobs: [:credential, :workflow, trigger: [:upstream_job]]
+    )
+    |> List.wrap()
     |> Workflows.to_project_space()
     |> Jason.encode!()
     |> Base.encode64()
@@ -113,9 +168,31 @@ defmodule LightningWeb.WorkflowLive do
      socket
      |> assign(
        active_menu_item: :projects,
-       encoded_project_space: encode_project_space(project),
        new_credential: false,
        builder_state: %{}
+     )}
+  end
+
+  @impl true
+  def handle_event("create-workflow", _, socket) do
+    {:ok, %Workflows.Workflow{id: workflow_id}} =
+      Workflows.create_workflow(%{project_id: socket.assigns.project.id})
+
+    {:noreply,
+     socket
+     |> assign(
+       page:
+         Workflows.get_workflows_for_query(socket.assigns.project)
+         |> Lightning.Repo.paginate(%{})
+     )
+     |> push_patch(
+       to:
+         Routes.project_workflow_path(
+           socket,
+           :show,
+           socket.assigns.project.id,
+           workflow_id
+         )
      )}
   end
 
@@ -124,14 +201,17 @@ defmodule LightningWeb.WorkflowLive do
   """
   @impl true
   def handle_info(
-        %Phoenix.Socket.Broadcast{event: "update", payload: _payload},
+        %Phoenix.Socket.Broadcast{
+          event: "update",
+          payload: %{workflow_id: workflow_id}
+        },
         socket
       ) do
+    workflow = Lightning.Workflows.get_workflow!(workflow_id)
+
     {:noreply,
      socket
-     |> assign(
-       encoded_project_space: encode_project_space(socket.assigns.project)
-     )}
+     |> assign(encoded_project_space: encode_project_space(workflow))}
   end
 
   # Update the builder state when an input dataclip is selected for a specific job
@@ -153,16 +233,31 @@ defmodule LightningWeb.WorkflowLive do
     {:noreply, apply_action(socket, socket.assigns.live_action, params)}
   end
 
-  defp apply_action(socket, :show, _params) do
+  # defp apply_action(socket, :show, _params) do
+  #   socket
+  #   |> assign(
+  #     active_menu_item: :overview,
+  #     page_title: socket.assigns.project.name
+  #   )
+  # end
+
+  defp apply_action(socket, :index, params) do
     socket
     |> assign(
       active_menu_item: :overview,
-      page_title: socket.assigns.project.name
+      page_title: socket.assigns.project.name,
+      show_canvas: false,
+      page:
+        Workflows.get_workflows_for_query(socket.assigns.project)
+        |> Lightning.Repo.paginate(params)
     )
   end
 
   defp apply_action(socket, :new_job, %{"upstream_id" => upstream_id}) do
     upstream_job = Lightning.Jobs.get_job!(upstream_id)
+
+    %Lightning.Jobs.Job{workflow: workflow} =
+      upstream_job |> Lightning.Repo.preload(:workflow)
 
     socket
     |> assign(
@@ -175,11 +270,18 @@ defmodule LightningWeb.WorkflowLive do
           "upstream_job_id" => upstream_job.id
         }
       },
+      current_workflow: workflow,
+      encoded_project_space: encode_project_space(workflow),
       page_title: socket.assigns.project.name
     )
   end
 
-  defp apply_action(socket, :new_job, %{"project_id" => project_id}) do
+  defp apply_action(socket, :new_job, %{
+         "project_id" => project_id,
+         "workflow_id" => workflow_id
+       }) do
+    workflow = Lightning.Workflows.get_workflow!(workflow_id)
+
     socket
     |> assign(
       active_menu_item: :overview,
@@ -187,7 +289,13 @@ defmodule LightningWeb.WorkflowLive do
       job_params: %{
         "trigger" => %{"type" => :webhook}
       },
-      workflow: Workflows.Workflow.new(%{project_id: project_id}),
+      current_workflow: workflow,
+      workflow:
+        Workflows.Workflow.changeset(workflow, %{
+          name: workflow.name,
+          project_id: project_id
+        }),
+      encoded_project_space: encode_project_space(workflow),
       page_title: socket.assigns.project.name
     )
   end
@@ -195,18 +303,54 @@ defmodule LightningWeb.WorkflowLive do
   defp apply_action(socket, :edit_job, %{"job_id" => job_id}) do
     job = Lightning.Jobs.get_job!(job_id)
 
+    %Lightning.Jobs.Job{workflow: workflow} =
+      job |> Lightning.Repo.preload(:workflow)
+
     socket
     |> assign(
       active_menu_item: :overview,
       job: job,
+      current_workflow: workflow,
+      encoded_project_space: encode_project_space(workflow),
       page_title: socket.assigns.project.name
     )
   end
 
-  defp apply_action(socket, :edit_workflow, %{"workflow_id" => workflow_id}) do
+  defp apply_action(socket, :edit_workflow, %{
+         "project_id" => project_id,
+         "workflow_id" => workflow_id
+       }) do
     workflow = Lightning.Workflows.get_workflow!(workflow_id)
 
     socket
-    |> assign(page_title: socket.assigns.project.name, workflow: workflow)
+    |> assign(
+      page_title: socket.assigns.project.name,
+      current_workflow: workflow,
+      encoded_project_space: encode_project_space(workflow),
+      workflow:
+        Workflows.Workflow.changeset(workflow, %{
+          name: workflow.name,
+          project_id: project_id
+        })
+    )
+  end
+
+  defp apply_action(socket, :show, %{"workflow_id" => workflow_id}) do
+    workflow =
+      Lightning.Workflows.get_workflow!(workflow_id)
+      |> Lightning.Repo.preload(
+        jobs: [:credential, :workflow, trigger: [:upstream_job]]
+      )
+
+    # we display the canvas only if workflow has jobs, otherwise we prompt the user to create a job
+    show_canvas = length(workflow.jobs) > 0
+
+    socket
+    |> assign(
+      page_title: socket.assigns.project.name,
+      current_workflow: workflow,
+      show_canvas: show_canvas,
+      encoded_project_space: encode_project_space(workflow)
+    )
   end
 end