+++ test/lightning_web/live/workflow_live_test.exs
@@ -8,18 +8,80 @@ defmodule LightningWeb.WorkflowLiveTest do
   setup :create_project_for_current_user
 
   import Lightning.JobsFixtures
+  import Lightning.WorkflowsFixtures
+
+  describe "index" do
+    test "lists all workflows for a project", %{
+      conn: conn,
+      project: project
+    } do
+      %{workflow: workflow_one} = workflow_job_fixture(project_id: project.id)
+      %{workflow: workflow_two} = workflow_job_fixture(project_id: project.id)
+
+      {:ok, view, html} =
+        live(conn, Routes.project_workflow_path(conn, :index, project.id))
+
+      assert html =~ "Create a workflow"
+
+      assert view
+             |> element(
+               "a[href='#{Routes.project_workflow_path(conn, :show, project.id, workflow_one.id)}']"
+             )
+             |> has_element?()
+
+      assert view
+             |> element(
+               "a[href='#{Routes.project_workflow_path(conn, :show, project.id, workflow_two.id)}']"
+             )
+             |> has_element?()
+    end
+  end
+
+  describe "create" do
+    test "Create empty workflow for a project", %{
+      conn: conn,
+      project: project
+    } do
+      {:ok, view, html} =
+        live(conn, Routes.project_workflow_path(conn, :index, project.id))
+
+      assert html =~ "Create a workflow"
+
+      assert view
+             |> element("button[phx-click='create-workflow']")
+             |> render_click() =~
+               "Add your first job"
+    end
+  end
 
   describe "show" do
-    setup %{project: project} do
-      %{job: job_fixture(project_id: project.id)}
+    test "renders prompt to create new job when workflow has no jobs", %{
+      conn: conn,
+      project: project
+    } do
+      workflow = workflow_fixture(name: "the workflow", project_id: project.id)
+
+      {:ok, _view, html} =
+        live(
+          conn,
+          Routes.project_workflow_path(conn, :show, project.id, workflow.id)
+        )
+
+      assert html =~ project.name
+      assert html =~ "Add your first job"
     end
 
     test "renders the workflow diagram", %{
       conn: conn,
       project: project
     } do
+      %{workflow: workflow} = workflow_job_fixture(project_id: project.id)
+
       {:ok, view, html} =
-        live(conn, Routes.project_workflow_path(conn, :show, project.id))
+        live(
+          conn,
+          Routes.project_workflow_path(conn, :show, project.id, workflow.id)
+        )
 
       assert html =~ project.name
 
@@ -37,7 +99,7 @@ defmodule LightningWeb.WorkflowLiveTest do
 
   describe "edit_job" do
     setup %{project: project} do
-      %{job: job_fixture(project_id: project.id)}
+      %{job: workflow_job_fixture(project_id: project.id)}
     end
 
     test "renders the job inspector", %{
@@ -48,7 +110,13 @@ defmodule LightningWeb.WorkflowLiveTest do
       {:ok, view, html} =
         live(
           conn,
-          Routes.project_workflow_path(conn, :edit_job, project.id, job.id)
+          Routes.project_workflow_path(
+            conn,
+            :edit_job,
+            project.id,
+            job.workflow_id,
+            job.id
+          )
         )
 
       assert html =~ project.name
@@ -71,13 +139,22 @@ defmodule LightningWeb.WorkflowLiveTest do
 
       view |> element("#job-form") |> render_submit()
 
-      assert_patch(view, Routes.project_workflow_path(conn, :show, project.id))
+      assert_patch(
+        view,
+        Routes.project_workflow_path(conn, :show, project.id, job.workflow_id)
+      )
 
       assert render(view) =~ "Job updated successfully"
 
       view
       |> render_patch(
-        Routes.project_workflow_path(conn, :edit_job, project.id, job.id)
+        Routes.project_workflow_path(
+          conn,
+          :edit_job,
+          project.id,
+          job.workflow_id,
+          job.id
+        )
       )
 
       assert has_element?(view, "#builder-#{job.id}")
@@ -91,7 +168,7 @@ defmodule LightningWeb.WorkflowLiveTest do
       conn: conn,
       project: project
     } do
-      upstream_job = job_fixture(project_id: project.id)
+      upstream_job = workflow_job_fixture(project_id: project.id)
 
       {:ok, view, html} =
         live(
@@ -100,6 +177,7 @@ defmodule LightningWeb.WorkflowLiveTest do
             conn,
             :new_job,
             project.id,
+            upstream_job.workflow_id,
             %{"upstream_id" => upstream_job.id}
           )
         )
@@ -149,7 +227,15 @@ defmodule LightningWeb.WorkflowLiveTest do
 
       view |> submit_form()
 
-      assert_patch(view, Routes.project_workflow_path(conn, :show, project.id))
+      assert_patch(
+        view,
+        Routes.project_workflow_path(
+          conn,
+          :show,
+          project.id,
+          upstream_job.workflow_id
+        )
+      )
 
       assert view |> encoded_project_space_matches(project)
     end
@@ -158,10 +244,12 @@ defmodule LightningWeb.WorkflowLiveTest do
       conn: conn,
       project: project
     } do
+      %{workflow: workflow} = workflow_job_fixture()
+
       {:ok, view, html} =
         live(
           conn,
-          Routes.project_workflow_path(conn, :new_job, project.id)
+          Routes.project_workflow_path(conn, :new_job, project.id, workflow.id)
         )
 
       assert html =~ project.name
@@ -180,7 +268,10 @@ defmodule LightningWeb.WorkflowLiveTest do
 
       view |> submit_form()
 
-      assert_patch(view, Routes.project_workflow_path(conn, :show, project.id))
+      assert_patch(
+        view,
+        Routes.project_workflow_path(conn, :show, project.id, workflow.id)
+      )
 
       assert view |> encoded_project_space_matches(project)
     end
@@ -188,7 +279,7 @@ defmodule LightningWeb.WorkflowLiveTest do
 
   describe "edit_workflow" do
     setup %{project: project} do
-      %{job: job_fixture(project_id: project.id)}
+      %{job: workflow_job_fixture(project_id: project.id)}
     end
 
     test "renders the workflow inspector", %{
@@ -217,7 +308,10 @@ defmodule LightningWeb.WorkflowLiveTest do
 
       view |> form("#workflow-form") |> render_submit()
 
-      assert_patch(view, Routes.project_workflow_path(conn, :show, project.id))
+      assert_patch(
+        view,
+        Routes.project_workflow_path(conn, :show, project.id, job.workflow_id)
+      )
 
       assert view |> encoded_project_space_matches(project)
     end
@@ -259,7 +353,7 @@ defmodule LightningWeb.WorkflowLiveTest do
     alias LightningWeb.JobLive.CronSetupComponent
 
     setup %{project: project} do
-      %{job: job_fixture(project_id: project.id)}
+      %{job: workflow_job_fixture(project_id: project.id)}
     end
 
     test "get_cron_data/1" do
@@ -358,12 +452,18 @@ defmodule LightningWeb.WorkflowLiveTest do
     test "cron_setup_component can create a new job with a default cron trigger",
          %{
            conn: conn,
-           project: project
+           project: project,
+           job: job
          } do
       {:ok, view, html} =
         live(
           conn,
-          Routes.project_workflow_path(conn, :new_job, project.id)
+          Routes.project_workflow_path(
+            conn,
+            :new_job,
+            project.id,
+            job.workflow_id
+          )
         )
 
       assert html =~ project.name
@@ -394,7 +494,11 @@ defmodule LightningWeb.WorkflowLiveTest do
              |> render_change()
 
       view |> submit_form() =~ "Job updated successfully"
-      assert_patch(view, Routes.project_workflow_path(conn, :show, project.id))
+
+      assert_patch(
+        view,
+        Routes.project_workflow_path(conn, :show, project.id, job.workflow_id)
+      )
 
       job =
         Lightning.Repo.get_by!(Lightning.Jobs.Job, %{name: "my job"})
@@ -416,7 +520,13 @@ defmodule LightningWeb.WorkflowLiveTest do
       {:ok, view, html} =
         live(
           conn,
-          Routes.project_workflow_path(conn, :edit_job, project.id, job.id)
+          Routes.project_workflow_path(
+            conn,
+            :edit_job,
+            project.id,
+            job.workflow_id,
+            job.id
+          )
         )
 
       assert job.trigger.type == :webhook
@@ -470,7 +580,13 @@ defmodule LightningWeb.WorkflowLiveTest do
       {:ok, view, html} =
         live(
           conn,
-          Routes.project_workflow_path(conn, :edit_job, project.id, job.id)
+          Routes.project_workflow_path(
+            conn,
+            :edit_job,
+            project.id,
+            job.workflow_id,
+            job.id
+          )
         )
 
       assert job.trigger.type == :webhook
@@ -518,7 +634,13 @@ defmodule LightningWeb.WorkflowLiveTest do
       {:ok, view, html} =
         live(
           conn,
-          Routes.project_workflow_path(conn, :edit_job, project.id, job.id)
+          Routes.project_workflow_path(
+            conn,
+            :edit_job,
+            project.id,
+            job.workflow_id,
+            job.id
+          )
         )
 
       assert job.trigger.type == :webhook
@@ -576,7 +698,13 @@ defmodule LightningWeb.WorkflowLiveTest do
       {:ok, view, html} =
         live(
           conn,
-          Routes.project_workflow_path(conn, :edit_job, project.id, job.id)
+          Routes.project_workflow_path(
+            conn,
+            :edit_job,
+            project.id,
+            job.workflow_id,
+            job.id
+          )
         )
 
       assert job.trigger.type == :webhook
@@ -616,7 +744,13 @@ defmodule LightningWeb.WorkflowLiveTest do
       {:ok, view, _html} =
         live(
           conn,
-          Routes.project_workflow_path(conn, :edit_job, project.id, job.id)
+          Routes.project_workflow_path(
+            conn,
+            :edit_job,
+            project.id,
+            job.workflow_id,
+            job.id
+          )
         )
 
       assert view |> has_element?("#minute", "10")